python/fastapi-sqlalchemy/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
PyMySQL==1.1.0
python-dotenv==1.0.0
email-validator==2.2.0
